move table publishing to its own file, reuse for bp dev
